e2e: make target URL resolution accessible to scripts
This allows the same logic to be used in the vegeta CLI script
that is used in the E2E testing that happens in the CI environment.
